Safari-611.1.10/LayoutTests/fast/dom/Document/editing-parameters.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Test that the parameters of the editing API on Document are mandatory.");

shouldThrow("document.execCommand()", "'TypeError: Not enough arguments'");
shouldThrow("document.queryCommandEnabled()", "'TypeError: Not enough arguments'");
shouldThrow("document.queryCommandIndeterm()", "'TypeError: Not enough arguments'");
shouldThrow("document.queryCommandState()", "'TypeError: Not enough arguments'");
shouldThrow("document.queryCommandSupported()", "'TypeError: Not enough arguments'");
shouldThrow("document.queryCommandValue()", "'TypeError: Not enough arguments'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
